requirements.txt
flask>=3.0
google-api-python-client>=2.0
pandas>=2.0
python-dotenv>=1.0
gunicorn>=21.0
gevent>=23.0
